+++ docs/source/changes.md
@@ -21,6 +21,11 @@ authors:
   In previous versions of the Pipeline, rejection was carried out before ICA
   and SSP. The exclude epochs from ICA fitting, use the new
   [`ica_reject`][config.ica_reject] setting.
+- We don't apply SSP by default anymore.
+  ({{ gh(315) }} by {{ authors.hoechenberger }})
+- The `use_ssp` and `use_ica` settings have been removed. Please use the
+  [`spatial_filter`][config.spatial_filter] setting instead.
+  ({{ gh(315) }} by {{ authors.hoechenberger }})
 
 ### Bug fixes